+++ examples/rbp/model.yaml
@@ -28,7 +28,7 @@ schema:
     targets:
         shape: (1, )
         descr: Predicted binding strength
-post_processing:
+postprocessing:
     - type: var_effect_prediction
       args:
           seq_input:
